# cmd/goapiexample/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\u003cbr\u003e\n\u003cbr\u003e\nThis API handles books operations.\n\u003cbr\u003e\n## Operations\n\u003cbr\u003e\nWith this API you can:\n\u003cbr\u003e\n* **Create book**\n* **Update book**\n* **Get all books**\n* **Get specific book with {id}**",
        "title": "Books DB application",
        "contact": {
            "name": "Mustafa Unal",
            "url": "https://github.com/bleakview/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/books": {
            "get": {
                "description": "Get all recorded books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Gets all books",
                "operationId": "GetAllBooks",
                "responses": {
                    "200": {
                        "description": "Book array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Book"
                            }
                        }
                    }
                }
            }
        },
        "/books/": {
            "post": {
                "description": "Create single book with given parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Book",
                "operationId": "CreateBook",
                "parameters": [
                    {
                        "description": "Book data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "description": "Gets a single book by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a book",
                "operationId": "GetBookById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "found book",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                }
            },
            "put": {
                "description": "Update book with give id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update book",
                "operationId": "UpdateBook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Book data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isbn": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "release_year": {
                    "type": "integer"
                }
            }
        },
        "model.BookRequest": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "isbn": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "release_year": {
                    "type": "integer"
                }
            }
        }
    }
}